ML 정보 반환 post -> alcohol 컨트롤러 변경 api 변경, alcolhole -> alcohol
